requirements.txt
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
boto3==1.17.41
botocore==1.20.41
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
distlib==0.3.1
docutils==0.15.2
dominate==2.6.0
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
flask-dynamo==0.1.2
flask-http-response==0.0.1.0
idna==2.10
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pipenv==2020.11.15
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
PyJWT==1.4.2
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.17.0
PyYAML==5.3.1
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
six==1.15.0
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
